适配mac/低内存 	modified:   0 zombie type/quick/main_quick_show.cpp 	new file:   "0 zombie type/\351\230\263\345\205\211\346\250\241\345\236\213/12-1 NE 1.5\350\212\261\344\272\224\345\205\253 hqyx/3\351\230\263\345\205\211\344\274\260\350\256\241-\347\240\264\351\230\265\347\216\207slow.cpp" 	modified:   "0 zombie type/\351\230\263\345\205\211\346\250\241\345\236\213/12-1 NE 1.5\350\212\261\344\272\224\345\205\253 hqyx/3\351\230\263\345\205\211\346\200\273\347\273\223.xlsx" 	modified:   inc/basic_setting.h

+++ inc/basic_setting.h
@@ -381,7 +381,7 @@ public:
         std::ofstream file(filename);
         std::sort(resu.begin(), resu.end());
         auto iter = resu.end()-1;
-        int upp = std::min(uint64_t(20),resu.size());
+        int upp = std::min(uint64_t(20),uint64_t(resu.size()));
         // 显示种子上限
         for(int i=0;i<upp;i++){
             file<<iter->metric;
